chore: Update file paths and imports for registerUser component after reorganizing file structure
