+++ src/app/dashboard/[email]/page.tsx
@@ -1,8 +1,8 @@
-import AllCourses from '@/app/dashboard/allCourses';
-import ClassesAndGrades from '@/app/dashboard/classesAndGrades';
-import UserSection from '@/app/dashboard/userSection/userSection';
+import AllCourses from '@/components/allCourses';
+import ClassesAndGrades from '@/components/classesAndGrades';
+import UserSection from '@/components/userSection/userSection';
 import { getUser } from '@workos-inc/authkit-nextjs'
-import { getAllCourses, getUserCourses, getUserGradedClasses } from '@/app/actions';
+import { getAllCourses, getUserCourses, getUserGradedClasses } from '@/actions/actions';
 
 
 export default async function Page() {
