+++ src/components/userSection/userSection.tsx
@@ -1,6 +1,6 @@
-import { UserCourse} from "@/app/types";
-import UserInfo from "@/app/dashboard/userSection/userInfo";
-import UserCourses from "@/app/dashboard/userSection/userCourses";
+import { UserCourse} from "@/types/types";
+import UserInfo from "@/components/userSection/userInfo";
+import UserCourses from "@/components/userSection/userCourses";
 export default function UserSection({ user, userCourses }: { user: any, userCourses: UserCourse[] }) {
     return (
         <div className='flex flex-col w-full md:w-96 gap-5 '>
